switch to zero-padded dates in `SINCE` query
without doing this, we get an error like:
```
Err(Bad("parsing time \"\" as \"_2-Jan-2006\": cannot parse \"\" as \"_2\""))
```

I think this is because the full query looks like:
```
Query("FROM \"info@itrevolution.com\" SINCE  4-Jan-2023")
```

Note the the two spaces after the `SINCE`. My guess is that the IMAP parser expects a single space and so is trying to read the space after that space as the beginning of a date, and then failing.